feat: iframe-based issues and MRs (#1330) (#1361)
